Remove references from ilias to outside
