+++ src/uploaders/ilias_folder.rs
@@ -3,16 +3,16 @@ use anyhow::{Ok, Result};
 use crate::{
     ilias::{
         client::IliasClient,
-        folder::{Folder, FolderElement},
+        folder::{Folder, FolderElement}, local_file::NamedLocalFile,
     },
     preselect_delete_setting::PreselectDeleteSetting,
 };
 
-use super::{file_data::FileData, upload_provider::UploadProvider};
+use super::upload_provider::UploadProvider;
 impl UploadProvider for Folder {
     type UploadedFile = FolderElement;
 
-    fn upload_files(&self, ilias_client: &IliasClient, file_data: &[FileData]) -> Result<()> {
+    fn upload_files(&self, ilias_client: &IliasClient, file_data: &[NamedLocalFile]) -> Result<()> {
         self.upload_files(ilias_client, file_data)
     }
 
@@ -47,7 +47,7 @@ impl UploadProvider for Folder {
     fn preselect_files<'a>(
         &self,
         preselect_setting: PreselectDeleteSetting,
-        upload_files: &[FileData],
+        upload_files: &[NamedLocalFile],
         existing_files: Vec<&'a Self::UploadedFile>,
     ) -> Vec<(&'a Self::UploadedFile, bool)> {
         existing_files
