+++ src/uploaders/mod.rs
@@ -1,5 +1,3 @@
 pub mod assignment_submission;
-pub mod file_data;
-pub mod file_with_filename;
 pub mod ilias_folder;
 pub mod upload_provider;
