+++ src/uploaders/upload_provider.rs
@@ -1,19 +1,17 @@
 use anyhow::Result;
 
-use crate::{ilias::client::IliasClient, preselect_delete_setting::PreselectDeleteSetting};
-
-use super::file_data::FileData;
+use crate::{ilias::{client::IliasClient, local_file::NamedLocalFile}, preselect_delete_setting::PreselectDeleteSetting};
 
 pub trait UploadProvider {
     type UploadedFile: ToString;
-    fn upload_files(&self, ilias_client: &IliasClient, file_data: &[FileData]) -> Result<()>;
+    fn upload_files(&self, ilias_client: &IliasClient, file_data: &[NamedLocalFile]) -> Result<()>;
     fn get_existing_files(&self) -> Vec<&Self::UploadedFile>;
     fn delete_files(&self, ilias_client: &IliasClient, files: &[&Self::UploadedFile])
         -> Result<()>;
     fn preselect_files<'a>(
         &self,
         preselect_setting: PreselectDeleteSetting,
-        upload_files: &[FileData],
+        upload_files: &[NamedLocalFile],
         existing_files: Vec<&'a Self::UploadedFile>,
     ) -> Vec<(&'a Self::UploadedFile, bool)>;
     /*fn select_files_to_delete<'a, I: Iterator<Item = FileData>>(
