+++ src/util.rs
@@ -1,25 +1,5 @@
 use clap::ValueEnum;
-use reqwest::Url;
 use serde::Deserialize;
-
-pub const ILIAS_URL: &str = "https://ilias.studium.kit.edu";
-
-pub trait Querypath {
-    fn get_querypath(&self) -> String;
-    fn set_querypath(&mut self, querypath: &str);
-}
-
-impl Querypath for Url {
-    fn get_querypath(&self) -> String {
-        format!("{}?{}", self.path(), self.query().unwrap_or(""))
-    }
-
-    fn set_querypath(&mut self, querypath: &str) {
-        let mut parts = querypath.split("?");
-        self.set_path(parts.next().unwrap());
-        self.set_query(parts.next());
-    }
-}
 
 #[derive(Debug, Deserialize, Clone, ValueEnum, PartialEq)]
 #[clap(rename_all = "kebab_case")]
